perf+fix: pre-tech-week audit pass — bundle wins, version fix, dashboard honesty, lint→0
Works the backlog from audit-reports/site-audit-2026-05-31/.

Performance
- /ingredients First Load JS 558→180 kB (-68%): the "use client" page imported
  the full ~2.6 MB ingredient catalog into the browser bundle. Split into a
  server component that projects the catalog to the ~20 fields the UI reads
  (src/lib/ingredients/slimIngredients.ts, server-only) + a client island
  (IngredientsExplorer.tsx) fed via props. SSR/SEO preserved — 1021 ingredients
  still prerendered.
- /profile First Load JS 268→215 kB: code-split the non-default-tab panels via
  next/dynamic.

Fixes
- Version mismatch: /api/health reported a stale 2.1.0 because
  npm_package_version is unset in Vercel's serverless runtime. next.config.js
  now inlines APP_VERSION from package.json (bumped 3.0.0→3.1.0); the health
  route and structured logger consume it.
- Admin dashboard honesty: dashboardPanelsService no longer substitutes
  fabricated values in its live path; CommercePanel / ElementalTraffic / the
  PractitionersCohort retention heatmap show honest live/"not wired" states;
  the FALLBACK_DATA seed is zeroed.
- lint:quick → 0 warnings: removed dead GlobalPopup.js; relaxed import/order for
  test files in the fast ESLint config (mirrors main, which ignores __tests__/).

Verified: bun run verify + bun run build green; browser QA on /ingredients
(search/filter/expand/pantry, 284 card photos render, no hydration errors). A
/code-review pass caught and fixed a dropped image_url field in the slim
projection (cards had lost their photos).
