+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "what-to-eat-next",
-  "version": "3.0.0",
+  "version": "3.1.0",
   "private": true,
   "type": "module",
   "engines": {
