+++ src/lib/mcp/__tests__/tools.test.ts
@@ -100,10 +100,10 @@ jest.mock("@/lib/mcp/rateLimit", () => ({
   __resetMcpRateLimit: jest.fn(),
 }));
 
-import { recordInvocation } from "@/lib/mcp/invocationLog";
 import { debitForTool } from "@/lib/mcp/auth";
-import { calculateNatalChart } from "@/services/natalChartService";
+import { recordInvocation } from "@/lib/mcp/invocationLog";
 import { invokeTool } from "@/lib/mcp/tools";
+import { calculateNatalChart } from "@/services/natalChartService";
 
 const mockedDebit = debitForTool as jest.MockedFunction<typeof debitForTool>;
 const mockedRecord = recordInvocation as jest.MockedFunction<
