+++ src/utils/elementalUtils.test.ts
@@ -1,7 +1,7 @@
 // src/__tests__/utils/elementalUtils.test.ts
 
-import { elementalUtils } from '@/utils/elementalUtils';
 import type { ElementalProperties } from '@/types/alchemy';
+import { elementalUtils } from '@/utils/elementalUtils';
 
 describe('elementalUtils', () => {
   let consoleWarnSpy: jest.SpyInstance;
